requirements.txt
environs==9.5.0
marshmallow==3.15.0
packaging==21.3
pyparsing==3.0.7
python-dotenv==0.19.2
pyzmq==22.3.0
websockets==10.2
